server/layer2/smartPedals/app/requirements.txt
Flask==3.1.0
pymongo==4.13.0
paho-mqtt==2.1.0
requests==2.32.4
webex_bot==1.0.4
twilio==9.7.0
